+++ src/test/java/com/reason/ide/comment/OclCommenterTest.java
@@ -9,24 +9,42 @@ import org.junit.runners.*;
 @RunWith(JUnit4.class)
 public class OclCommenterTest extends ORBasePlatformTestCase {
     @Test
-    public void testLineCommenter() {
-        configureCode("A.ml", "comment<caret>");
+    public void test_line_commenter() {
+        configureCode("A.ml", "  comment<caret>");
         CommentByLineCommentAction action = new CommentByLineCommentAction();
 
         action.actionPerformedImpl(getProject(), myFixture.getEditor());
-        myFixture.checkResult("(*comment*)");
+        myFixture.checkResult("(*  comment *)");
 
         action.actionPerformedImpl(getProject(), myFixture.getEditor());
-        myFixture.checkResult("comment");
+        myFixture.checkResult("  comment ");
     }
 
     @Test
-    public void testLineUncommenter() {
-        configureCode("A.ml", "(* comment *)<caret>");
+    public void test_line_uncommenter() {
+        configureCode("A.ml", "  (* comment *)<caret>");
         CommentByLineCommentAction action = new CommentByLineCommentAction();
 
         action.actionPerformedImpl(getProject(), myFixture.getEditor());
-        myFixture.checkResult(" comment ");
+        myFixture.checkResult("   comment ");
+    }
+
+    @Test
+    public void test_line_uncommenter_02() {
+        configureCode("A.ml", "  (*comment*)<caret>");
+        CommentByLineCommentAction action = new CommentByLineCommentAction();
+
+        action.actionPerformedImpl(getProject(), myFixture.getEditor());
+        myFixture.checkResult("  comment");
+    }
+
+    @Test
+    public void test_line_uncommenter_03() {
+        configureCode("A.ml", "\ntest\n(*    comment    *)    <caret>");
+        CommentByLineCommentAction action = new CommentByLineCommentAction();
+
+        action.actionPerformedImpl(getProject(), myFixture.getEditor());
+        myFixture.checkResult("\ntest\n    comment        ");
     }
 
     /*--
@@ -37,7 +55,7 @@ public class OclCommenterTest extends ORBasePlatformTestCase {
     <caret>
     --*/
     @Test
-    public void testGH_27() {
+    public void test_GH_27() {
         configureCode("A.ml", "<selection>(*\n  x (* y *)\n*)</selection>\n<caret>");
 
         CommentByBlockCommentAction action = new CommentByBlockCommentAction();
@@ -55,7 +73,7 @@ public class OclCommenterTest extends ORBasePlatformTestCase {
                            *)
     --*/
     @Test
-    public void testGH_27b() {
+    public void test_GH_27b() {
         configureCode("A.ml", "<selection>(* x *)\ny\n(* z *)\n</selection><caret>");
 
         CommentByBlockCommentAction commentAction = new CommentByBlockCommentAction();
@@ -65,12 +83,33 @@ public class OclCommenterTest extends ORBasePlatformTestCase {
     }
 
     @Test
-    public void testGH_27c() {
+    public void test_GH_27c() {
         configureCode("A.ml", "<selection>(* x *)\n(* z *)\n</selection><caret>");
 
         CommentByBlockCommentAction commentAction = new CommentByBlockCommentAction();
         commentAction.actionPerformedImpl(getProject(), myFixture.getEditor());
 
         myFixture.checkResult("(*\n(* x *)\n(* z *)\n*)\n");
+        assertEquals(19, myFixture.getCaretOffset());
+    }
+
+    @Test
+    public void test_GH_319() {
+        configureCode("A.ml", "line with (<caret>");
+
+        CommentByLineCommentAction a = new CommentByLineCommentAction();
+        a.actionPerformedImpl(getProject(), myFixture.getEditor());
+
+        myFixture.checkResult("(* line with ( *)");
+    }
+
+    @Test
+    public void test_GH_319_b() {
+        configureCode("A.ml", " line with ( <caret>");
+
+        CommentByLineCommentAction a = new CommentByLineCommentAction();
+        a.actionPerformedImpl(getProject(), myFixture.getEditor());
+
+        myFixture.checkResult("(* line with ( *)");
     }
 }