+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 ## Unreleased
 
 - :bug: [#350](https://github.com/giraud/reasonml-idea-plugin/issues/350) Use declaration documentation if no local documentation
+- :bug: [#319](https://github.com/giraud/reasonml-idea-plugin/issues/319) "Comment with line comment" action incorrect for lines ending with "("
 - :bug: [#174](https://github.com/giraud/reasonml-idea-plugin/issues/174) Compiler output listener needs a change for ocaml >= 4.09
 
 ## 0.110 - 2023/03/01